+++ storage/key_format.go
@@ -1,9 +1,10 @@
 package storage
 
 import (
-	"encoding/binary"
 	"fmt"
 	"strings"
+
+	"github.com/hyperledger/burrow/binary"
 )
 
 const (
@@ -81,8 +82,8 @@ func (kf *KeyFormat) KeyBytes(segments ...[]byte) ([]byte, error) {
 			offset--
 		default:
 			if len(s) != l {
-				return nil, fmt.Errorf("the segment '0x%X' provided to KeyFormat.KeyBytes() does not have required "+
-					"%d bytes required by layout for segment %d", s, l, i)
+				return nil, fmt.Errorf("the key segment '%X' (length %d) provided to KeyFormat.KeyBytes() "+
+					"does not the have required %d bytes required by layout for segment %d", s, len(s), l, i)
 			}
 			n += l
 			// Big endian so pad on left if not given the full width for this segment
@@ -235,9 +236,9 @@ func scan(a interface{}, value []byte) {
 		// Ignore - allows for omitted values
 	case *int64:
 		// Negative values will be mapped correctly when read in as uint64 and then type converted
-		*v = int64(binary.BigEndian.Uint64(value))
+		*v = binary.GetInt64(value)
 	case *uint64:
-		*v = binary.BigEndian.Uint64(value)
+		*v = binary.GetUint64(value)
 	case *[]byte:
 		*v = value
 	case *string:
@@ -250,14 +251,14 @@ func scan(a interface{}, value []byte) {
 func format(a interface{}) []byte {
 	switch v := a.(type) {
 	case uint64:
-		return formatUint64(v)
+		return binary.Uint64Bytes(v)
 	case int64:
-		return formatUint64(uint64(v))
+		return binary.Uint64Bytes(uint64(v))
 	// Provide formatting from int,uint as a convenience to avoid casting arguments
 	case uint:
-		return formatUint64(uint64(v))
+		return binary.Uint64Bytes(uint64(v))
 	case int:
-		return formatUint64(uint64(v))
+		return binary.Uint64Bytes(uint64(v))
 	case []byte:
 		return v
 	case ByteSlicable:
@@ -267,12 +268,6 @@ func format(a interface{}) []byte {
 	default:
 		panic(fmt.Errorf("KeyFormat format() does not support formatting value of type %T: %v", a, a))
 	}
-}
-
-func formatUint64(v uint64) []byte {
-	bs := make([]byte, 8)
-	binary.BigEndian.PutUint64(bs, v)
-	return bs
 }
 
 // MustKeyFormat for panicking early when a KeyFormat does not parse